# shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ChatMod Mobile API
  version: 0.1.0
  description: Backend contract for the phone-hosted YouTube Live moderation bot.
servers:
  - url: http://localhost:4100
paths:
...
  /backups:
    get:
      summary: List cloud backups for the current account
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account-scoped backup list
  /backups/moderation-profile:
    post:
      summary: Store a moderation profile backup
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Backup stored
  /backups/settings:
    post:
      summary: Store a validated command and timer settings backup
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Settings backup stored
  /backups/{id}:
    delete:
      summary: Delete a cloud backup for the current account
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Backup deleted
  /backups/{id}/restore:
    post:
      summary: Restore a settings backup into commands and timers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings backup restored
        "403":
          description: Current plan command, timed-message, or channel profile limit reached
...
  /integrations/discord/test:
    post:
      summary: Send a test Discord alert for a configured profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - profileId
              properties:
                profileId:
                  type: string
      responses:
        "200":
          description: Test alert result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscordAlertResult"
        "403":
          description: Discord alerts require Pro or Creator plan
        "409":
          description: Discord webhook is not configured or enabled
...
  /admin/support/users:
    get:
      summary: Admin support lookup by device id
      security:
        - adminApiKey: []
      parameters:
        - in: query
          name: deviceId
          required: true
          schema:
            type: string
        - in: query
          name: installId
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Device-scoped support snapshot with entitlement, subscription, diagnostics, and API errors
        "401":
          description: Missing or invalid admin API key
  /admin/support/devices/{deviceId}:
    get:
      summary: Admin device/session support lookup
      security:
        - adminApiKey: []
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device-scoped support snapshot
        "401":
          description: Missing or invalid admin API key
  /admin/support/subscriptions/{deviceId}:
    get:
      summary: Admin subscription and entitlement lookup
      security:
        - adminApiKey: []
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device-scoped entitlement/subscription snapshot
        "401":
          description: Missing or invalid admin API key
  /admin/support/entitlements/manual-adjust:
    post:
      summary: Admin manual entitlement adjustment
      security:
        - adminApiKey: []
      responses:
        "200":
          description: Entitlement updated and support audit event recorded
        "401":
          description: Missing or invalid admin API key
  /admin/support/tickets/metadata:
    post:
      summary: Admin support ticket metadata event
      security:
        - adminApiKey: []
      responses:
        "201":
          description: Support ticket metadata stored as a support event
        "401":
          description: Missing or invalid admin API key
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminApiKey:
      type: apiKey
      in: header
      name: x-admin-api-key
  schemas:
...
    DiscordAlertResult:
      type: object
      required:
        - sent
        - profileId
      properties:
        sent:
          type: boolean
        skippedReason:
          type:
            - string
            - "null"
          enum:
            - not_configured
            - disabled
            - moderation_alerts_disabled
            - runtime_alerts_disabled
            - null
        sentAt:
          type:
            - string
            - "null"
          format: date-time
        profileId:
          type: string
...